frontend/requirements.txt
streamlit==1.29.0
requests==2.32.3
google-auth==2.39.0